Add files via upload
Problem Statement-The Department of Symbiartic Engineering offers both arts and science co
urses. You have a list of students
entering the second year.  Each student record consists of a nine-letter
roll number (like 15SY10049,
15SY60D04, 15SY92P15), an arts marks (out of 100, an integer), an
d a science marks (again an integer, out
of 100). A student record typically contains many other fields, but you do
not need to worry about those for
the sake of this assignment. Assume that the list is sorted with respect to the roll
numbers.
An array of structures stores the list. According to the policy of the Departme
nt, the list is presented to you
in a read-only array, and you are not allowed to modify the array or copy
any record from the array to any
other location in the memory. You may only visit any record, read any field, do
comparisons on this value,
and if needed print it. You are forbidden to use any record or any of its c
omponents on the right side of
any assignme