Add files via upload
Let T be a binary tree with each node storing a flag called EOS (see below why) an d two child pointers L and R. Some nodes inT, including all the leaves, have the EOS flag set. Imagine that every left-child link is labeled by 0, and every right-child link by 1. We say that T stores a binary string α ∈ {0,1} ∗ if there is a node v with the EOS flag set such that the unique path from the root to v is labeled by the symbols of α. Here, EOS stands for the end of a string stored in T