Add files via upload
Heaps- You are given two arrays A and B storing m and n integers. Both A and Bare unsorted. An AB-sum is a sum of the form A[i] +B[j] for some (valid) indices i and j. You are also given a positive integer k satisfying k 6m /log m and 6 nlog n . Your task is to find out the k smallest-AB-sums. Notice that it is not important which i and which j correspond to an AB-sum. In particular, if you permute A and B these indices change, but the k smallest AB -sums remain the same. However, if the same AB-sum is realized by two or more different index pairs (for the same permutations of A and B), these sums are to be reported separately